+++ packages/terra-consumer-nav/src/components/user-profile/ProfileLinks.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import Arrange from 'terra-arrange';
-import Button from 'terra-button';
 import IconChevronDown from 'terra-icon/lib/icon/IconChevronDown';
 import IconChevronUp from 'terra-icon/lib/icon/IconChevronUp';
 import Toggler from 'terra-toggle';
@@ -53,13 +52,13 @@ class ProfileLinks extends React.Component {
         const toggleIcon = isOpen ? <IconChevronUp /> : <IconChevronDown />;
         return (
           <div key={linkItem.text} className={cx(isOpen && 'open')}>
-            <Button onClick={() => { this.handleToggle(index); }} className={cx('link', 'toggler-wrapper')}>
+            <button onClick={() => { this.handleToggle(index); }} className={cx('link', 'toggler-wrapper')}>
               <Arrange
                 align="stretch"
                 fill={<div>{linkItem.text}</div>}
                 fitEnd={<div>{toggleIcon}</div>}
               />
-            </Button>
+            </button>
             <Toggler isOpen={isOpen} isAnimated className={cx('toggler')}>
               {linkItem.subItems.map(subItem => (<ProfileLink key={subItem.text} {...subItem} />))}
             </Toggler>